Untested draft of Shenzen IO solitaire solver
I hacked together a first crack at finding solutions to the solitaire
mini game in Shenzen IO. I haven't created any tests, just got the rough
idea to compile without warnings.

The basic idea is to have a game state with "free" cells, "stacks" (of
cards) and "finished" cells. Any single card can go in a free cell.
Cards of the same suit can be stack in ascending order in the "finished"
cells (the goal of Solitaire) and the cards to work with are dealt out
into "stacks" (or piles) in shuffled order. We take the state, figure out
all the possible "moves" (a recorded move and the resulting state),
test them for solving the game or else generate new moves based on
existing states. Each new stack that is generated within a state because
of a move is tested against all previous stacks: we don't want to
recreate an exact stack we've already seen. This should prevent simple
cycles but there will probably be inefficiencies because of silly moves.
If it works, but is too inefficient, I can order the moves so that the
kinds of moves most likely to advance the game happen first. The entire
solution set is exhausted.

The game is complicated by "dragon" sets of four cards that can be
removed ("slain") to an empty free cell, removing that free cell from
play. In order to remove/slay a dragon all of the cards in the set must
be on the bottom of a stack or in a free cell.

Instead of the traditional red and black colours of solitaire this game
has green as well. Colours still have to alternate: red -> green/black,
green -> red/black or black -> red/green.

Possible moves are:
- any card at the bottom of a stack to a free cell
- any card at the bottom of a stack to a finished cell
  - if the card is a 1 and the finished cell is empty, or
  - if the card is 1 more than the top card in the finished cell, and
    is the same suit as the top card in the finished cell
- any sub-stack of cards on a stack can be moved to another stack
  - if the target stack is empty, or
  - if the cards in the sub-stack are in descending order, and
    the cards in the sub-stack alternate in colour, and
    the top card of the target stack is a different colour than the
    bottom card of the source sub-stack, and
    the top card of the target stack is 1 more than the bottom card
    of the source stack.
- any card in a free cell can be moved to the bottom of a stack if
  all the rules are followed for moving sub-stacks; i.e. the card
  in the free cell is considered a sub-stack.