Use V, not KV for map value; show moves to card
I was setting the free cell values to {{free, N}, foo} instead of just
to foo.

When displaying moves I can just show the target card we're moving to.
The rest of the stack is irrelevant except maybe for double checking
that the state looks correct. That is, if the target stack doesn't match
the game board then something is wrong in the model of the game.