Wrong Turn
Turns out I missed a card the entire time I was playing: the ...
Poppy? D'oh, this whole time I thought it was a rose. Maybe it isn't a
poppy either, I dunno. Anyways, I started by adding moves from a stack
to free and from a stack to empty: however, YOU ONLY NEED TO MOVE THE
POPPY DIRECTLY TO THE POPPY PILE! Yes, there's a dedicated poppy pile
that's just slightly separate from the other finished piles and has,
suprise, a poppy outline in it.

So, I'm checking this unfinished code in before I turf it all and write
a simple function to move the poppy to the poppy space, just in case I'm
wrong.