The poppy has a dedicated space
If the poppy is visible it can always be moved. I'm actually not sure
what the point of it is, since as soon as it's visible it's
automatically moved. My guess is that it's just a filler card to bring
the totall up to 40: 3 stacks of 9 cards, 3 sets of 4 dragons and 1
poppy card.