Handle poppies, fix substack generation
Fix the starting state in the solve function:
- need free cells
- need moves list

Put in a limiter since the code never stops. I've stubbed out a test for
is_solved/1 in case that's broken.

Add a poppy move: any visible poppy is basically just deleted since it
has it's own cell and is never heard from again.

Wire the "rest of source stack" value: I didn't intend to do this but I
got confused and ended up adding it. Weird, I know: how do you
accidentally add a feature. I dunno, I started working on one problem
and ended up adding a different feature.

Handle empty "other stack": when moving to a different stack, that stack
can be empty. In that case, we can't get the top card.

Ditch is_poppy_visible because it's a pretty simple list comprehension.

Handle dragons and the poppy in is_valid_stack_move/1 to see if we can
even move a stack. Also check for a poppy as the top of the target stack
(can't stack on top of a poppy). Really, I should just delete the poppy
as the first move and not even bother calculating other moves.

Added a test to solve the game board with a real board example from
Shenzen IO but it just runs forever, so I'm guessing is_solved/1 might
be busted. I'm also adding a "mini" board so that it doesn't take as
long.

Fixed the substack generation code to not generate and empty substack.