Make it work (but not all the tests)
I've got this working with a sample game but I want to get it checked in
before I worry about fixing up the broken tests.

Add the solution to the solve test. I can't try it because that game was
closed. I'll try another game. I need to work on an easier way to enter
the cards.

Fixed and tested cards_to_finish_moves/1:
- finish cells need to be in the starting state
- finish cells should start with ['empty'] (atom in list)
- differentiate between empty atom and num+suit tuple
  - move the match out of the function head and into a case statement
- finish cells have suits, not numbers

Add some code to print out lists of moves once they get to a certain
length.

Depth-first instead of breadth-first: when we get a list of working
moves, we'll put those at the front of the list and work on them first,
instead of trying all the other starting moves first. Let's follow the
original path to its conclusion.

Re-order movement generation to put most-desirable moves first. The most
desirable move is to get rid of the poppy, because that's free and easy;
second move is to put a card on its finished pile, third is to slay a
dragon and get it out of the way, etc.