Formatting
I thought I had an issue but it was because I had entered a duplicate
card.

I tried one of the solutions in the game even though it was only a
partial solution and I managed to win. There's a couple things I need to
do:
- Always start with the largest movable substack: I was moving a lot of
stacks back and forth
- Move from free to finish! <- this is a kicker